chore(tests): `tests/utils/config` package
