+++ .gitignore
@@ -23,7 +23,6 @@ test_data
 trie_putandget_failed_test_data_*
 tmp
 
-tests/utils/config*
 tests/utils/genesis*
 
 # node_modules used by polkadot.js/api tests
