+++ tests/rpc/rpc_01-system_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ChainSafe/gossamer/dot/rpc/modules"
 	libutils "github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/ChainSafe/gossamer/tests/utils"
+	"github.com/ChainSafe/gossamer/tests/utils/config"
 	"github.com/stretchr/testify/require"
 )
 
@@ -91,7 +92,7 @@ func TestSystemRPC(t *testing.T) {
 
 	t.Log("starting gossamer...")
 	genesisPath := libutils.GetGssmrGenesisRawPathTest(t)
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitializeAndStartNodes(t, 3, genesisPath, config)
 
 	//use only first server for tests
