+++ tests/stress/stress_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/ChainSafe/gossamer/lib/common"
 	libutils "github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/ChainSafe/gossamer/tests/utils"
+	"github.com/ChainSafe/gossamer/tests/utils/config"
 	"github.com/ChainSafe/gossamer/tests/utils/rpc"
 )
 
@@ -55,7 +56,7 @@ func TestMain(m *testing.M) {
 
 func TestRestartNode(t *testing.T) {
 	numNodes := 1
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitNodes(numNodes, config)
 	require.NoError(t, err)
 
@@ -79,14 +80,14 @@ func TestSync_SingleBlockProducer(t *testing.T) {
 	// start block producing node first
 	basePath := t.TempDir()
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	configNoGrandpa := utils.CreateConfigNoGrandpa(t)
+	configNoGrandpa := config.CreateNoGrandpa(t)
 	node, err := utils.RunGossamer(t, numNodes-1,
 		basePath,
 		genesisPath, configNoGrandpa,
 		false, true)
 	require.NoError(t, err)
 
-	configNoAuthority := utils.CreateConfigNotAuthority(t)
+	configNoAuthority := config.CreateNotAuthority(t)
 
 	// wait and start rest of nodes - if they all start at the same time the first round usually doesn't complete since
 	// all nodes vote for different blocks.
@@ -127,7 +128,7 @@ func TestSync_SingleBlockProducer(t *testing.T) {
 func TestSync_Basic(t *testing.T) {
 	genesisPath := libutils.GetGssmrGenesisRawPathTest(t)
 
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitializeAndStartNodes(t, 3, genesisPath, config)
 	require.NoError(t, err)
 
@@ -150,7 +151,7 @@ func TestSync_MultipleEpoch(t *testing.T) {
 	utils.Logger.Patch(log.SetLevel(log.Info))
 
 	// wait and start rest of nodes - if they all start at the same time the first round usually doesn't complete since
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitializeAndStartNodes(t, numNodes, genesisPath, config)
 	require.NoError(t, err)
 
@@ -203,16 +204,16 @@ func TestSync_SingleSyncingNode(t *testing.T) {
 	// start block producing node
 	blockProducingNodebasePath := t.TempDir()
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	config := utils.CreateDefaultConfig(t)
+	configPath := config.CreateDefault(t)
 	alice, err := utils.RunGossamer(t, 0,
 		blockProducingNodebasePath, genesisPath,
-		config, false, true)
+		configPath, false, true)
 	require.NoError(t, err)
 	time.Sleep(time.Second * 15)
 
 	// start syncing node
 	syncingNodeBasePath := t.TempDir()
-	configPath := utils.CreateConfigNoBabe(t)
+	configPath = config.CreateNoBabe(t)
 	bob, err := utils.RunGossamer(t, 1,
 		syncingNodeBasePath, genesisPath,
 		configPath, false, false)
@@ -246,7 +247,7 @@ func TestSync_Bench(t *testing.T) {
 	// start block producing node
 	blockProducingNodebasePath := t.TempDir()
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	configNoGrandpa := utils.CreateConfigNoGrandpa(t)
+	configNoGrandpa := config.CreateNoGrandpa(t)
 	alice, err := utils.RunGossamer(t, 0,
 		blockProducingNodebasePath,
 		genesisPath, configNoGrandpa,
@@ -279,7 +280,7 @@ func TestSync_Bench(t *testing.T) {
 
 	// start syncing node
 	syncingNodeBasePath := t.TempDir()
-	configNoAuthority := utils.CreateConfigNotAuthority(t)
+	configNoAuthority := config.CreateNotAuthority(t)
 	bob, err := utils.RunGossamer(t, 1,
 		syncingNodeBasePath, genesisPath,
 		configNoAuthority, false, true)
@@ -346,16 +347,16 @@ func TestSync_Restart(t *testing.T) {
 	// start block producing node first
 	blockProducingNodeBasePath := t.TempDir()
 	genesisPath := libutils.GetGssmrGenesisRawPathTest(t)
-	config := utils.CreateDefaultConfig(t)
+	configPath := config.CreateDefault(t)
 	node, err := utils.RunGossamer(t, numNodes-1,
 		blockProducingNodeBasePath,
-		genesisPath, config,
+		genesisPath, configPath,
 		false, true)
 	require.NoError(t, err)
 
 	// wait and start rest of nodes
 	time.Sleep(time.Second * 5)
-	configPath := utils.CreateConfigNoBabe(t)
+	configPath = config.CreateNoBabe(t)
 	nodes, err := utils.InitializeAndStartNodes(t, numNodes-1, genesisPath, configPath)
 	require.NoError(t, err)
 	nodes = append(nodes, node)
@@ -415,14 +416,14 @@ func TestSync_SubmitExtrinsic(t *testing.T) {
 	// start block producing node first
 	blockProducingNodeBasePath := t.TempDir()
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	configNoGrandpa := utils.CreateConfigNoGrandpa(t)
+	configNoGrandpa := config.CreateNoGrandpa(t)
 	node, err := utils.RunGossamer(t, 0,
 		blockProducingNodeBasePath, genesisPath,
 		configNoGrandpa, false, true)
 	require.NoError(t, err)
 	nodes := []utils.Node{node}
 
-	configNoAuthority := utils.CreateConfigNotAuthority(t)
+	configNoAuthority := config.CreateNotAuthority(t)
 
 	// Start rest of nodes
 	basePath2 := t.TempDir()
@@ -583,7 +584,7 @@ func Test_SubmitAndWatchExtrinsic(t *testing.T) {
 	// start block producing node first
 	blockProducingNodeBasePath := t.TempDir()
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	configNoGrandpa := utils.CreateConfigNoGrandpa(t)
+	configNoGrandpa := config.CreateNoGrandpa(t)
 	node, err := utils.RunGossamer(t, 0,
 		blockProducingNodeBasePath,
 		genesisPath, configNoGrandpa, true, true)
@@ -763,7 +764,7 @@ func TestStress_SecondarySlotProduction(t *testing.T) {
 	const numNodes = 2
 	for _, c := range testcases {
 		t.Run(c.description, func(t *testing.T) {
-			config := utils.CreateDefaultConfig(t)
+			config := config.CreateDefault(t)
 			nodes, err := utils.InitializeAndStartNodes(t, numNodes, c.genesis, config)
 			require.NoError(t, err)
 			defer utils.StopNodes(t, nodes)
