+++ tests/utils/framework.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 	"testing"
 
+	"github.com/ChainSafe/gossamer/tests/utils/config"
 	"github.com/ChainSafe/gossamer/tests/utils/rpc"
 	scribble "github.com/nanobox-io/golang-scribble"
 )
@@ -24,7 +25,7 @@ type Framework struct {
 // InitFramework creates given quanity of nodes
 func InitFramework(t *testing.T, qtyNodes int) (*Framework, error) {
 	f := &Framework{}
-	configPath := CreateDefaultConfig(t)
+	configPath := config.CreateDefault(t)
 
 	nodes, err := InitNodes(qtyNodes, configPath)
 	if err != nil {
